// src/generated/unitRegistry.ts
export const UNIT_REGISTRY_VERSION = "curve-units-v1";
export const UNIT_REGISTRY_SHA256 = "28f8a5e5ccbbf8ac4c2e8cdfd0d12928c54c13229e161cb15daecf83897f3443";
export const UNIT_REGISTRY_FAMILIES = [
  {
    "family": "GR",
    "canonicalUnit": "gAPI",
    "quantityKind": "gamma_ray",
    "aliases": [
      "GR",
      "GRN",
      "GRD",
      "CGR",
      "SGR",
      "GRGC",
      "GRKT"
    ]
  },
  {
    "family": "SP",
    "canonicalUnit": "mV",
    "quantityKind": "electric_potential",
    "aliases": [
      "SP",
      "SPC",
      "SPR"
    ]
  },
  {
    "family": "CALI",
    "canonicalUnit": "in",
    "quantityKind": "length",
    "aliases": [
      "CALI",
      "CAL",
      "CALS",
      "CALX",
      "CALY",
      "HCAL",
      "LCAL",
      "DCAL",
      "HORD"
    ]
  },
  {
    "family": "BS",
    "canonicalUnit": "in",
    "quantityKind": "length",
    "aliases": [
      "BS",
      "BITSIZE",
      "BIT"
    ]
  },
  {
    "family": "RHOB",
    "canonicalUnit": "g/cc",
    "quantityKind": "bulk_density",
    "aliases": [
      "RHOB",
      "RHOZ",
      "RHOBED",
      "DEN",
      "ZDEN",
      "ROBB",
      "SBD2"
    ]
  },
  {
    "family": "DRHO",
    "canonicalUnit": "g/cc",
    "quantityKind": "bulk_density",
    "aliases": [
      "DRHO",
      "HDRA",
      "ZCOR",
      "DCOR"
    ]
  },
  {
    "family": "PEF",
    "canonicalUnit": "b/e",
    "quantityKind": "photoelectric_factor",
    "aliases": [
      "PEF",
      "PE",
      "PEFZ",
      "PEB",
      "PDPE"
    ]
  },
  {
    "family": "NPHI",
    "canonicalUnit": "v/v",
    "quantityKind": "fraction",
    "aliases": [
      "NPHI",
      "TNPH",
      "NPHIED",
      "NPHI_LS",
      "NPOR",
      "NEUT",
      "APLC",
      "FPLC",
      "SNP",
      "HNPO",
      "FSTP"
    ]
  },
  {
    "family": "DT",
    "canonicalUnit": "us/ft",
    "quantityKind": "slowness",
    "aliases": [
      "DT",
      "DTC",
      "DTCO",
      "AC",
      "DT24",
      "DTP",
      "DTCOMP"
    ]
  },
  {
    "family": "DTS",
    "canonicalUnit": "us/ft",
    "quantityKind": "slowness",
    "aliases": [
      "DTS",
      "DTSM",
      "DTSH",
      "DTSHEAR",
      "DT_S"
    ]
  },
  {
    "family": "TEMP",
    "canonicalUnit": "DEGC",
    "quantityKind": "temperature",
    "aliases": [
      "FTEMP"
    ]
  },
  {
    "family": "RES_DEEP",
    "canonicalUnit": "ohm.m",
    "quantityKind": "resistivity",
    "aliases": [
      "RES_DEEP",
      "RESD",
      "RT",
      "RDEEP",
      "RDEP",
      "DRES",
      "ILD",
      "LLD",
      "AT90",
      "AHT90",
      "RLA5",
      "ATR",
      "BDAV",
      "RING",
      "PSR"
    ]
  },
  {
    "family": "RES_MED",
    "canonicalUnit": "ohm.m",
    "quantityKind": "resistivity",
    "aliases": [
      "RES_MED",
      "RESM",
      "RMED",
      "ILM",
      "LLM",
      "AT30",
      "AHT30",
      "RLA3"
    ]
  },
  {
    "family": "RES_SHAL",
    "canonicalUnit": "ohm.m",
    "quantityKind": "resistivity",
    "aliases": [
      "RES_SHAL",
      "RESS",
      "RSHAL",
      "SFL",
      "SFLU",
      "LL8",
      "SN",
      "AT10",
      "AHT10",
      "RLA1",
      "R25P",
      "BSAV"
    ]
  },
  {
    "family": "RXO",
    "canonicalUnit": "ohm.m",
    "quantityKind": "resistivity",
    "aliases": [
      "RXO",
      "RXOZ",
      "MSFL",
      "RMLL"
    ]
  }
] as const;
export const UNIT_REGISTRY_UNITS = [
  {
    "token": "gAPI",
    "quantityKind": "gamma_ray",
    "canonicalUnit": "gAPI"
  },
  {
    "token": "GAPI",
    "quantityKind": "gamma_ray",
    "canonicalUnit": "gAPI"
  },
  {
    "token": "mV",
    "quantityKind": "electric_potential",
    "canonicalUnit": "mV"
  },
  {
    "token": "m",
    "quantityKind": "length",
    "canonicalUnit": "m"
  },
  {
    "token": "M",
    "quantityKind": "length",
    "canonicalUnit": "m"
  },
  {
    "token": "mm",
    "quantityKind": "length",
    "canonicalUnit": "mm"
  },
  {
    "token": "MM",
    "quantityKind": "length",
    "canonicalUnit": "mm"
  },
  {
    "token": "cm",
    "quantityKind": "length",
    "canonicalUnit": "cm"
  },
  {
    "token": "CM",
    "quantityKind": "length",
    "canonicalUnit": "cm"
  },
  {
    "token": "in",
    "quantityKind": "length",
    "canonicalUnit": "in"
  },
  {
    "token": "IN",
    "quantityKind": "length",
    "canonicalUnit": "in"
  },
  {
    "token": "g/cc",
    "quantityKind": "bulk_density",
    "canonicalUnit": "g/cc"
  },
  {
    "token": "G/CC",
    "quantityKind": "bulk_density",
    "canonicalUnit": "g/cc"
  },
  {
    "token": "G/C3",
    "quantityKind": "bulk_density",
    "canonicalUnit": "g/cc"
  },
  {
    "token": "kg/m3",
    "quantityKind": "bulk_density",
    "canonicalUnit": "kg/m3"
  },
  {
    "token": "KG/M3",
    "quantityKind": "bulk_density",
    "canonicalUnit": "kg/m3"
  },
  {
    "token": "k/m3",
    "quantityKind": "bulk_density",
    "canonicalUnit": "kg/m3"
  },
  {
    "token": "K/M3",
    "quantityKind": "bulk_density",
    "canonicalUnit": "kg/m3"
  },
  {
    "token": "b/e",
    "quantityKind": "photoelectric_factor",
    "canonicalUnit": "b/e"
  },
  {
    "token": "B/E",
    "quantityKind": "photoelectric_factor",
    "canonicalUnit": "b/e"
  },
  {
    "token": "v/v",
    "quantityKind": "fraction",
    "canonicalUnit": "v/v"
  },
  {
    "token": "V/V",
    "quantityKind": "fraction",
    "canonicalUnit": "v/v"
  },
  {
    "token": "pu",
    "quantityKind": "fraction",
    "canonicalUnit": "pu"
  },
  {
    "token": "PU",
    "quantityKind": "fraction",
    "canonicalUnit": "pu"
  },
  {
    "token": "%",
    "quantityKind": "fraction",
    "canonicalUnit": "%"
  },
  {
    "token": "p.u.",
    "quantityKind": "fraction",
    "canonicalUnit": "pu"
  },
  {
    "token": "P.U.",
    "quantityKind": "fraction",
    "canonicalUnit": "pu"
  },
  {
    "token": "us/m",
    "quantityKind": "slowness",
    "canonicalUnit": "us/m"
  },
  {
    "token": "US/M",
    "quantityKind": "slowness",
    "canonicalUnit": "us/m"
  },
  {
    "token": "usec/m",
    "quantityKind": "slowness",
    "canonicalUnit": "us/m"
  },
  {
    "token": "USEC/M",
    "quantityKind": "slowness",
    "canonicalUnit": "us/m"
  },
  {
    "token": "us/ft",
    "quantityKind": "slowness",
    "canonicalUnit": "us/ft"
  },
  {
    "token": "US/FT",
    "quantityKind": "slowness",
    "canonicalUnit": "us/ft"
  },
  {
    "token": "degF",
    "quantityKind": "temperature",
    "canonicalUnit": "degF"
  },
  {
    "token": "DEGF",
    "quantityKind": "temperature",
    "canonicalUnit": "degF"
  },
  {
    "token": "degC",
    "quantityKind": "temperature",
    "canonicalUnit": "degC"
  },
  {
    "token": "DEGC",
    "quantityKind": "temperature",
    "canonicalUnit": "degC"
  },
  {
    "token": "ohm.m",
    "quantityKind": "resistivity",
    "canonicalUnit": "ohm.m"
  },
  {
    "token": "OHM.M",
    "quantityKind": "resistivity",
    "canonicalUnit": "ohm.m"
  },
  {
    "token": "meq/L",
    "quantityKind": "charge_per_volume",
    "canonicalUnit": "meq/L"
  },
  {
    "token": "MEQ/L",
    "quantityKind": "charge_per_volume",
    "canonicalUnit": "meq/L"
  },
  {
    "token": "meq/mL",
    "quantityKind": "charge_per_volume",
    "canonicalUnit": "meq/mL"
  },
  {
    "token": "MEQ/ML",
    "quantityKind": "charge_per_volume",
    "canonicalUnit": "meq/mL"
  },
  {
    "token": "md",
    "quantityKind": "permeability",
    "canonicalUnit": "mD"
  },
  {
    "token": "mD",
    "quantityKind": "permeability",
    "canonicalUnit": "mD"
  }
] as const;
export const UNIT_REGISTRY_RULES = [
  {
    families: ["CALI","BS"],
    fromUnit: "mm",
    toUnit: "in",
    factor: 1.0 / 25.4,
    offset: 0.0,
    derivation: "1 in = 25.4 mm exactly; mm -> in divides by 25.4",
    automatic: true,
  },
  {
    families: ["CALI","BS"],
    fromUnit: "cm",
    toUnit: "in",
    factor: 1.0 / 2.54,
    offset: 0.0,
    derivation: "1 in = 2.54 cm exactly; cm -> in divides by 2.54",
    automatic: true,
  },
  {
    families: ["DT","DTS"],
    fromUnit: "us/m",
    toUnit: "us/ft",
    factor: 0.3048,
    offset: 0.0,
    derivation: "1 international ft = 0.3048 m; (us/m) x (m/ft) = us/ft",
    automatic: true,
  },
  {
    families: ["DT","DTS"],
    fromUnit: "usec/m",
    toUnit: "us/ft",
    factor: 0.3048,
    offset: 0.0,
    derivation: "1 usec = 1 us and 1 international ft = 0.3048 m; factor = 0.3048",
    automatic: true,
  },
  {
    families: ["RHOB","DRHO"],
    fromUnit: "kg/m3",
    toUnit: "g/cc",
    factor: 0.001,
    offset: 0.0,
    derivation: "1 g/cc = 1000 kg/m3; kg/m3 -> g/cc divides by 1000",
    automatic: true,
  },
  {
    families: ["NPHI"],
    fromUnit: "pu",
    toUnit: "v/v",
    factor: 0.01,
    offset: 0.0,
    derivation: "1 porosity unit = 1 percent = 0.01 v/v",
    automatic: true,
  },
  {
    families: ["NPHI"],
    fromUnit: "%",
    toUnit: "v/v",
    factor: 0.01,
    offset: 0.0,
    derivation: "1 percent = 1/100 = 0.01 v/v",
    automatic: true,
  },
  {
    families: ["NPHI"],
    fromUnit: "p.u.",
    toUnit: "v/v",
    factor: 0.01,
    offset: 0.0,
    derivation: "p.u. denotes porosity percent; 1 percent = 0.01 v/v",
    automatic: true,
  },
  {
    families: ["TEMP"],
    fromUnit: "DEGF",
    toUnit: "DEGC",
    factor: 1.0 / 1.8,
    offset: -32.0,
    derivation: "T42 fixes degC = (degF - 32) / 1.8; factor = 1/1.8 and offset = -32 degF",
    automatic: true,
  },
  {
    families: ["QV"],
    fromUnit: "MEQ/L",
    toUnit: "meq/mL",
    factor: 1.0e-3,
    offset: 0.0,
    derivation: "1 L = 10^3 mL; meq/L -> meq/mL therefore multiplies by 10^-3",
    automatic: false,
  }
] as const;
export const UNIT_REGISTRY_POPULATION = {
  families: UNIT_REGISTRY_FAMILIES.length,
  aliases: UNIT_REGISTRY_FAMILIES.reduce((sum, family) => sum + family.aliases.length, 0),
  units: UNIT_REGISTRY_UNITS.length,
  rules: UNIT_REGISTRY_RULES.length,
} as const;